--- direct-test-output.log
@@ -1,29 +1,0 @@
-Starting test...
-Starting direct currency selection test...
-Launching browser...
-Logs are being written to: /Users/nv/repos/0x4007/travel-stipend/logs/direct-test-2025-03-19T05-28-16.877Z.log
-Browser launched
-Navigating to Google Flights...
-[2025-03-19T05:28:18.824Z] [INFO] Navigating to Google Flights via Google homepage
-[2025-03-19T05:28:18.825Z] [INFO] Navigating to Google homepage
-[2025-03-19T05:28:20.238Z] [INFO] Waiting for Google homepage to stabilize
-[2025-03-19T05:28:23.242Z] [INFO] Searching for 'Google Flights'
-[2025-03-19T05:28:23.259Z] [WARN] Could not find search input, trying alternative approach
-[2025-03-19T05:28:23.259Z] [INFO] Trying direct navigation to Google Flights
-[2025-03-19T05:28:23.877Z] [INFO] Waiting for Google Flights page to load
-[2025-03-19T05:28:28.882Z] [INFO] Google Flights page loaded
-[2025-03-19T05:28:28.885Z] [INFO] Page loaded: Google Flights - Find Cheap Flight Options & Track Prices (https://www.google.com/travel/flights)
-[2025-03-19T05:28:28.886Z] [INFO] Successfully navigated to Google Flights
-Navigated to Google Flights
-Finding and clicking currency button...
-Currency button found and clicked: true
-Selecting USD in currency dialog...
-[2025-03-19T05:28:30.438Z] [INFO] Approach 1: Using Puppeteer's built-in selectors
-[2025-03-19T05:28:30.459Z] [INFO] Approach 2: Using page.evaluate for DOM manipulation
-[2025-03-19T05:28:30.567Z] [INFO] Approach 3: Using XPath selectors
-[2025-03-19T05:28:30.584Z] [INFO] Found and clicked USD element with XPath: //div[@role='dialog']//span[contains(text(), 'USD') or contains(text(), 'US Dollar')]
-USD selected in currency dialog: true
-Test completed!
-Closing browser...
-Browser closed
-Test function completed
