+++ knip.json
@@ -1,0 +1,31 @@
+{
+  "entry": [
+    "build/index.ts",
+    "build/esbuild-server.ts",
+    "build/esbuild-build.ts",
+    "src/travel-stipend-calculator.ts",
+    "src/historical-stipend-calculator.ts",
+    ".github/empty-string-checker.ts",
+    "rename-to-kebab-case.ts"
+  ],
+  "project": [
+    "src/**/*.ts"
+  ],
+  "ignore": [
+    "src/types/config.ts",
+    "**/__mocks__/**",
+    "**/__fixtures__/**",
+    "eslint.config.mjs",
+    "fixtures/split-historical.ts",
+    "src/utils/google-flights-scraper/**/*.ts"
+  ],
+  "ignoreExportsUsedInFile": true,
+  "ignoreDependencies": [
+    "@mswjs/data"
+  ],
+  "exclude": [
+    "files",
+    "devDependencies"
+  ],
+  "eslint": true
+}
