--- test-dashboard.md
@@ -1,20 +1,0 @@
-# Test Dashboard
-
-| :clock10: Start time   | :hourglass: Duration |
-| ---------------------- | -------------------: |
-| 3/19/2025, 12:06:24 PM |              2.609 s |
-
-|             | :white_check_mark: Passed | :x: Failed | :construction: Todo | :white_circle: Total |
-| ----------- | ------------------------: | ---------: | ------------------: | -------------------: |
-| Test Suites |                         1 |          0 |                   - |                    1 |
-| Tests       |                         4 |          0 |                   0 |                    4 |
-
-## tests/currency-selection.test.ts [[link](https://github.com/0x4007/travel-stipend/blob/9304e62b9c8aa1a4a5dbdcea34bb57283f0a7ea1/tests/currency-selection.test.ts)]
-
-4 passed, 0 failed, 0 todo, done in 2.548 s
-
-- :white_check_mark: GoogleFlightsScraper Currency Selection
-  - :white_check_mark: should not attempt to change currency if already USD
-  - :white_check_mark: should successfully change currency to USD
-  - :white_check_mark: should handle failure to find currency button
-  - :white_check_mark: should handle failure to select USD
