+++ tests/date-selection/select-date-in-calendar.ts
