+++ tests/date-selection/test-date-selection/click-found-search-button.ts
