+++ tests/date-selection/test-date-selection/clickable-element-info.ts
