+++ tests/date-selection/test-date-selection/find-and-click-search-button.ts
