+++ tests/date-selection/test-date-selection/log-clickable-elements.ts
