+++ tests/date-selection/test-date-selection/take-result-screenshot.ts
