+++ tests/date-selection/test-date-selection/try-java-script-click-approach.ts
