+++ tests/date-selection/try-java-script-click.ts
