+++ tests/date-selection/try-multiple-click-approaches.ts
