+++ tests/specific-destination-input/select-dropdown-item.ts
