+++ tests/specific-destination-input/verify-selection.ts
