+++ tests/test-currency-selection.ts
@@ -1,4 +1,4 @@
-import { GoogleFlightsScraper } from "../utils/google-flights-scraper";
+import { GoogleFlightsScraper } from "../src/utils/google-flights-scraper";
 
 // Set up process listeners for better debugging
 process.on("unhandledRejection", (reason, promise) => {
