+++ tests/test-destination-input.ts
